requirements.txt
uvicorn==0.20.0
fastapi==0.91.0
pytorch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
pytorch-cuda=11.6
numpy==1.21.5
opencv-python==4.7.0
scipy==1.8.0
gdown==4.6.3
transformers==4.25.1
tensorboardx==2.5.1
